Reorder/resize varaibles in Staking20Upgradeable.
- uint256 stakingTokenDecimals => uint16 stakingTokenDecimals
- uint256 rewardTokenDecimals => uint16 rewardTokenDecimals
- Move `stakersArray` below the fixed size variables.

+++ contracts/staking/TokenStake.sol
@@ -66,10 +66,10 @@ contract TokenStake is
         require(_rewardToken != _stakingToken, "Reward Token and Staking Token can't be same.");
         rewardToken = _rewardToken;
 
-        uint256 _stakingTokenDecimals = _stakingToken == CurrencyTransferLib.NATIVE_TOKEN
+        uint16 _stakingTokenDecimals = _stakingToken == CurrencyTransferLib.NATIVE_TOKEN
             ? 18
             : IERC20Metadata(_stakingToken).decimals();
-        uint256 _rewardTokenDecimals = _rewardToken == CurrencyTransferLib.NATIVE_TOKEN
+        uint16 _rewardTokenDecimals = _rewardToken == CurrencyTransferLib.NATIVE_TOKEN
             ? 18
             : IERC20Metadata(_rewardToken).decimals();
